# Conflicts:
#	frontend/src/services/utils.js

+++ backend/src/main/java/coworking/digitalBooking/Repository/CoworkingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoworkingRepository extends JpaRepository<Coworking,Integer> {
+public interface CoworkingRepository extends JpaRepository<Coworking,Long> {
 
     Coworking findByName(String name);
